Modernize & clean up issue contribution
* move duplicate code into common
* move code from random files into the contributions
* use classes instead of global scope (move to ProductService instead of product directly)
* use common naming (Native for electron, Web for web)
