+++ src/vs/workbench/contrib/issue/browser/issue.contribution.ts
@@ -1,0 +1,29 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as nls from 'vs/nls';
+import { CommandsRegistry } from 'vs/platform/commands/common/commands';
+import { InstantiationType, registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
+import { IProductService } from 'vs/platform/product/common/productService';
+import { Registry } from 'vs/platform/registry/common/platform';
+import { Extensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
+import { WebIssueService } from 'vs/workbench/contrib/issue/browser/issueService';
+import { IWorkbenchIssueService } from 'vs/workbench/services/issue/common/issue';
+import { BaseIssueContribution } from 'vs/workbench/contrib/issue/common/issue.contribution';
+
+class WebIssueContribution extends BaseIssueContribution {
+	constructor(@IProductService productService: IProductService) {
+		super(productService);
+	}
+}
+
+Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkbenchContribution(WebIssueContribution, LifecyclePhase.Restored);
+
+registerSingleton(IWorkbenchIssueService, WebIssueService, InstantiationType.Delayed);
+
+CommandsRegistry.registerCommand('_issues.getSystemStatus', (accessor) => {
+	return nls.localize('statusUnsupported', "The --status argument is not yet supported in browsers.");
+});
