--- src/vs/workbench/contrib/issue/common/commands.ts
@@ -1,14 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export const OpenIssueReporterActionId = 'workbench.action.openIssueReporter';
-
-export const OpenIssueReporterApiCommandId = 'vscode.openIssueReporter';
-
-export interface OpenIssueReporterArgs {
-	readonly extensionId?: string;
-	readonly issueTitle?: string;
-	readonly issueBody?: string;
-}
