+++ src/vs/workbench/services/issue/electron-sandbox/issueService.ts
@@ -23,7 +23,7 @@ import { IWorkspaceTrustManagementService } from 'vs/platform/workspace/common/w
 import { IIntegrityService } from 'vs/workbench/services/integrity/common/integrity';
 import { process } from 'vs/base/parts/sandbox/electron-sandbox/globals';
 
-export class WorkbenchIssueService implements IWorkbenchIssueService {
+export class NativeIssueService implements IWorkbenchIssueService {
 	declare readonly _serviceBrand: undefined;
 
 	constructor(
